seperated all chess related source files + this allow a more convinient usage of the MCTS for a non-chess related game type without using Stockfish as a third party dependency
