+++ engine/src/agents/mctsagent.cpp
@@ -27,9 +27,7 @@
 #include "mctsagent.h"
 #include "../evalinfo.h"
 #include "movegen.h"
-#include "inputrepresentation.h"
-#include "outputrepresentation.h"
-#include "constants.h"
+#include "../constants.h"
 #include "../util/blazeutil.h"
 #include "uci.h"
 #include "../manager/treemanager.h"
