+++ engine/src/searchthread.cpp
@@ -32,8 +32,6 @@
 
 #include <stdlib.h>
 #include <climits>
-#include "inputrepresentation.h"
-#include "outputrepresentation.h"
 #include "util/blazeutil.h"
 #include "uci.h"
