+++ engine/src/stateobj.h
@@ -29,11 +29,17 @@
 #ifndef STATEOBJ_H
 #define STATEOBJ_H
 
+#include <unordered_map>
 #include "state.h"
+#include <blaze/Math.h>
+using blaze::StaticVector;
+using blaze::DynamicVector;
+
 #ifdef MODE_POMMERMAN
 #include "pommermanstate.h"
 #else
-#include "boardstate.h"
+#include "chess_related/boardstate.h"
+#include "chess_related/outputrepresentation.h"
 #endif
 
 #ifdef MODE_POMMERMAN
@@ -43,5 +49,44 @@
 #endif
 
 std::string action_to_uci(Action action, bool is960);
+
+#ifndef MODE_POMMERMAN
+namespace Constants {
+inline void init(bool isPolicyMap) {
+    init_policy_constants(isPolicyMap);
+}
+}
+#endif
+
+/**
+ * @brief get_probs_of_move_list Returns an array in which entry relates to the probability for the given move list.
+                                 Its assumed that the moves in the move list are legal and shouldn't be mirrored.
+ * @param batchIdx Index to use in policyProb when extracting the probabilities for all legal moves
+ * @param policyProb Policy array from the neural net prediction
+ * @param legalMoves List of legal moves for a specific board position
+ * @param lastLegalMove Pointer to the last legal move
+ * @param sideToMove Determine if it's white's or black's turn to move
+ * @param normalize True, if the probability should be normalized
+ * @param selectPolicyFromPlane Sets if the policy is encoded in policy map representation
+ * @return policyProbSmall - A hybrid blaze vector which stores the probabilities for the given move list
+ */
+void get_probs_of_move_list(const size_t batchIdx, const float* policyProb, const std::vector<Action>& legalMoves, SideToMove sideToMove,
+                            bool normalize, DynamicVector<float> &policyProbSmall, bool selectPolicyFromPlane);
+
+/**
+ * @brief get_policy_data_batch Returns the pointer of the batch for the policy predictions
+ * @param batchIdx Batch index for the current predicion
+ * @param policyProb All policy predicitons from the batch
+ * @param isPolicyMap Sets if the policy is encoded in policy map representation
+ * @return Starting pointer for predictions of the current batch
+ */
+const float*  get_policy_data_batch(const size_t batchIdx, const float* policyProb, bool isPolicyMap);
+
+/**
+ * @brief get_current_move_lookup Returns the look-up table to use depending on the side to move
+ * @param sideToMove Current side to move
+ * @return Returns either MOVE_LOOK_UP or MOVE_LOOK_UP_MIRRORED
+ */
+std::unordered_map<Action, size_t, std::hash<int>>& get_current_move_lookup(SideToMove sideToMove);
 
 #endif // STATEOBJ_H
